refactor: Event 함수들 분리, productUtils 생성하여 분리
