+++ src/basic/utils/index.js
@@ -1,2 +1,3 @@
 export * from "./commonUtils";
 export * from "./cartUtils";
+export * from "./productUtils";
